Cambiar Integer por Long

+++ alcance-backend/gestion_presupuesto/comando/comando-aplicacion/src/main/java/com/ceiba/core/comando/ComandoPresupuesto.java
@@ -14,5 +14,5 @@ public class ComandoPresupuesto {
     private Double valor;
     private Double reservaEstabilizacion;
     private String observaciones;
-    private Integer idProyecto;
+    private Long idProyecto;
 }
